refactor(search): extract history/schedule tools to _HistoryToolsMixin
